Brings the Apple Silicon port onto main as a reviewable series:

  import the port from the scratch checkout it was stranded in
  add CMake presets and port documentation
  add the repeatable Infinity-vs-FAISS benchmark mechanism
  delete the abandoned Python attestation/campaign layer
  delete the unreferenced ctpl_overlay_regression suite
  delete orphan experiment harness sources
  remove Apple experiment regions from src/CMakeLists.txt
  resolve experiment macros in the HNSW engine sources
  drop execution-witness hooks from the benchmark bridge
  make the benchmark campaign binding optional
  record the complete SIFT1M baseline and iso-recall bracket

The port builds natively on arm64-apple-darwin with no Docker or Rosetta. The
engine reduction was verified bit-identical (same graph sha256, edge count and
recall on a deterministic single-threaded build).

First real-scale measurement, SIFT1M, M=32, 12 threads, from-source Accelerate
FAISS 1.15.0: at equal efConstruction Infinity builds 1.10x-1.31x faster than
FAISS; at matched recall it is 1.03x-1.18x slower. See docs/apple_silicon/BASELINE.md.

Net diff over upstream: 144 files, ~12.3k insertions -- down from the ~101k the
raw salvage carried, with the experiment scaffolding, attestation machinery and
rejected optimizations removed.

+++ src/parser/type/complex/tensor_array_type.cpp
@@ -21,7 +21,7 @@ import std.compat;
 #endif
 
 namespace infinity {
-std::string TensorArrayType::TensorArray2String(const std::vector<std::pair<std::span<const char>, uint64_t>> &raw_data,
+std::string TensorArrayType::TensorArray2String(const std::vector<std::pair<std::span<const char>, size_t>> &raw_data,
                                                 EmbeddingDataType type,
                                                 size_t embedding_dimension) {
     std::ostringstream oss;
